applying windows change to linux as well, gotta use common denominator types

+++ plugins/DebuggerCore/unix/osx/PlatformEvent.cpp
@@ -223,7 +223,7 @@ edb::tid_t PlatformEvent::thread() const {
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-int PlatformEvent::code() const {
+int64_t PlatformEvent::code() const {
 	if (stopped()) {
 		return WSTOPSIG(status);
 	}
